added SimpleHolochain js class
