+++ ui/package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@holochain/client": "^0.18.0-rc.0",
+    "@holochain/simple-holochain": "file:../lib",
     "@holochain-open-dev/utils": "0.400.0-dev.4",
     "@lit/context": "^1.1.3",
     "@lit/task": "^1.0.1",
