+++ ui/src/elements/edit-thing.ts
@@ -1,16 +1,17 @@
-import { ActionHash, AgentPubKey, AppClient, DnaHash, EntryHash, HolochainError, Record } from "@holochain/client";
+import { ActionHash, HolochainError, Record } from "@holochain/client";
 import { consume } from "@lit/context";
 import { decode } from "@msgpack/msgpack";
 import { html, LitElement } from "lit";
 import { customElement, property, state } from "lit/decorators.js";
 
-import { clientContext } from "../../contexts";
-import { Thing } from "./types";
+import { simpleHolochainContext } from "../contexts";
+import { Thing } from "@holochain/simple-holochain";
+import { SimpleHolochain } from "@holochain/simple-holochain";
 
-@customElement("edit-thing")
-export class EditThing extends LitElement {
-  @consume({ context: clientContext })
-  client!: AppClient;
+@customElement("edit-post")
+export class EditPost extends LitElement {
+  @consume({ context: simpleHolochainContext })
+  client!: SimpleHolochain;
 
   @property({
     hasChanged: (newVal: ActionHash, oldVal: ActionHash) => newVal?.toString() !== oldVal?.toString(),
@@ -44,35 +45,9 @@ export class EditThing extends LitElement {
     this._content = this.currentThing.content;
   }
 
-  async updateThing() {
-    const thing: Thing = {
-      content: this._content!,
-    };
-
+  async updatePost() {
     try {
-      const updateRecord: Record = await this.client.callZome({
-        cap_secret: null,
-        role_name: "generic_dna",
-        zome_name: "generic_zome",
-        fn_name: "update_thing",
-        payload: {
-          original_thing_hash: this.originalThingHash,
-          previous_thing_hash: this.currentRecord.signed_action.hashed.hash,
-          updated_thing: thing,
-        },
-      });
-
-      this.dispatchEvent(
-        new CustomEvent("thing-updated", {
-          composed: true,
-          bubbles: true,
-          detail: {
-            originalThingHash: this.originalThingHash,
-            previousThingHash: this.currentRecord.signed_action.hashed.hash,
-            updatedThingHash: updateRecord.signed_action.hashed.hash,
-          },
-        }),
-      );
+      await this.client.updateThing(this.originalThingHash, this._content);
     } catch (e) {
       alert((e as HolochainError).message);
     }
@@ -105,7 +80,7 @@ export class EditThing extends LitElement {
           >
             Cancel
           </button>
-          <button .disabled=${!this.isThingValid()} @click=${() => this.updateThing()}>
+          <button .disabled=${!this.isThingValid()} @click=${() => this.updatePost()}>
             Save
           </button>
         </div>
