--- ui/src/generic_dna/generic_zome/types.ts
@@ -1,46 +1,0 @@
-import {
-  ActionHash,
-  AgentPubKey,
-  Create,
-  CreateLink,
-  Delete,
-  DeleteLink,
-  DnaHash,
-  EntryHash,
-  ExternalHash,
-  Record,
-  SignedActionHashed,
-  Update,
-} from "@holochain/client";
-
-export type GenericZomeSignal = {
-  type: "EntryCreated";
-  action: SignedActionHashed<Create>;
-  app_entry: EntryTypes;
-} | {
-  type: "EntryUpdated";
-  action: SignedActionHashed<Update>;
-  app_entry: EntryTypes;
-  original_app_entry: EntryTypes;
-} | {
-  type: "EntryDeleted";
-  action: SignedActionHashed<Delete>;
-  original_app_entry: EntryTypes;
-} | {
-  type: "LinkCreated";
-  action: SignedActionHashed<CreateLink>;
-  link_type: string;
-} | {
-  type: "LinkDeleted";
-  action: SignedActionHashed<DeleteLink>;
-  link_type: string;
-};
-
-/* dprint-ignore-start */
-export type EntryTypes =
- | ({  type: 'Thing'; } & Thing);
-/* dprint-ignore-end */
-
-export interface Thing {
-  content: string;
-}
